Backend/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain_groq
dotenv
ddgs
sqlalchemy
psycopg2-binary
langchain-community
langchain-core
passlib[bcrypt]
python-jose
pydantic[email]
bcrypt==4.0.1
python-multipart